requirements.txt
streamlit
requests
beautifulsoup4
spacy
python-dotenv
google-generativeai
python-docx 
PyPDF2
seaborn
reportlab
pandas
matplotlib
langdetect
easyocr
